remove: icons dependencies and all imports
