+++ web-client/package.json
@@ -42,7 +42,6 @@
     "immer": "^10.1.1",
     "ky": "^1.7.5",
     "lodash": "^4.17.21",
-    "lucide-react": "^0.471.2",
     "moment": "^2.30.1",
     "next": "15.1.4",
     "next-auth": "5.0.0-beta.25",
@@ -52,7 +51,6 @@
     "react-calendar": "^5.1.0",
     "react-dom": "^19.0.0",
     "react-hook-form": "^7.54.2",
-    "react-icons": "^5.5.0",
     "tailwind-merge": "^2.6.0",
     "tailwindcss-animate": "^1.0.7",
     "zod": "^3.24.2",
