+++ web-client/src/components/ui/breadcrumb.tsx
@@ -1,6 +1,5 @@
 import { cn } from '@/lib/utils';
 import { Slot } from '@radix-ui/react-slot';
-import { ChevronRight, MoreHorizontal } from 'lucide-react';
 import * as React from 'react';
 
 const Breadcrumb = React.forwardRef<
@@ -82,7 +81,7 @@ const BreadcrumbSeparator = ({
     className={cn('[&>svg]:w-3.5 [&>svg]:h-3.5', className)}
     {...props}
   >
-    {children ?? <ChevronRight />}
+    {children ?? null}
   </li>
 );
 BreadcrumbSeparator.displayName = 'BreadcrumbSeparator';
@@ -97,7 +96,6 @@ const BreadcrumbEllipsis = ({
     className={cn('flex h-9 w-9 items-center justify-center', className)}
     {...props}
   >
-    <MoreHorizontal className='h-4 w-4' />
     <span className='sr-only'>More</span>
   </span>
 );
