+++ web-client/src/components/ui/dialog.tsx
@@ -1,8 +1,8 @@
 'use client';
 
+import Icon, { IconName } from '@/components/utils/icon';
 import { cn } from '@/lib/utils';
 import * as DialogPrimitive from '@radix-ui/react-dialog';
-import { X } from 'lucide-react';
 import * as React from 'react';
 
 const Dialog = DialogPrimitive.Root;
@@ -47,7 +47,7 @@ const DialogContent = React.forwardRef<
       {children}
       {!hideCloseButton && (
         <DialogPrimitive.Close className='absolute right-4 top-4 rounded-sm opacity-70 ring-offset-background transition-opacity hover:opacity-100 focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2 disabled:pointer-events-none data-[state=open]:bg-accent data-[state=open]:text-muted-foreground'>
-          <X className='h-4 w-4' />
+          <Icon name={IconName.CLOSE} />
           <span className='sr-only'>Close</span>
         </DialogPrimitive.Close>
       )}
