+++ web-client/src/components/ui/sheet.tsx
@@ -1,9 +1,9 @@
 'use client';
 
+import Icon, { IconName } from '@/components/utils/icon';
 import { cn } from '@/lib/utils';
 import * as SheetPrimitive from '@radix-ui/react-dialog';
 import { type VariantProps, cva } from 'class-variance-authority';
-import { X } from 'lucide-react';
 import * as React from 'react';
 
 const Sheet = SheetPrimitive.Root;
@@ -64,7 +64,7 @@ const SheetContent = React.forwardRef<
       {...props}
     >
       <SheetPrimitive.Close className='absolute right-4 top-4 rounded-sm opacity-70 ring-offset-background transition-opacity hover:opacity-100 focus:outline-none focus:ring-2 focus:ring-ring focus:ring-offset-2 disabled:pointer-events-none data-[state=open]:bg-secondary'>
-        <X className='h-4 w-4' />
+        <Icon name={IconName.CLOSE} />
         <span className='sr-only'>Close</span>
       </SheetPrimitive.Close>
       {children}
