+++ web-client/src/components/utils/icon/Icon.stories.ts
@@ -1,5 +1,4 @@
-import Icon from './Icon';
-import { IconName } from '@/constants/icons';
+import Icon, { IconName } from './Icon';
 import type { Meta, StoryObj } from '@storybook/react';
 
 const meta = {
