+++ web-client/src/components/utils/icon/index.tsx
@@ -1,0 +1,1 @@
+export { default, IconName } from './Icon';
