+++ web-client/src/components/utils/theme.tsx
@@ -1,7 +1,6 @@
 'use client';
 
 import { Switch } from '@/components/ui/switch';
-import { Moon, Sun } from 'lucide-react';
 import { useTheme } from 'next-themes';
 
 export const ToggleDarkModeButton = () => {
@@ -9,11 +8,7 @@ export const ToggleDarkModeButton = () => {
 
   return (
     <div className='flex justify-center items-center'>
-      {resolvedTheme === 'dark' ? (
-        <Moon className='h-[1.2rem] w-[1.2rem] rotate-90 scale-0 transition-all fill-black dark:rotate-0 dark:scale-100' />
-      ) : (
-        <Sun className='h-[1.2rem] w-[1.2rem] rotate-0 scale-100 transition-all dark:-rotate-90 dark:scale-0' />
-      )}
+      {resolvedTheme === 'dark' ? null : null}
 
       <Switch
         checked={resolvedTheme === 'dark'}
