--- web-client/src/constants/icons.ts
@@ -1,27 +1,0 @@
-export enum IconName {
-  HOME = 'home',
-
-  CLOSE = 'close',
-
-  UP = 'up',
-  DOWN = 'down',
-  LEFT = 'left',
-  RIGHT = 'right',
-
-  THEME = 'theme',
-
-  LIGHT_MODE = 'light-mode',
-  DARK_MODE = 'dark-mode',
-
-  MENU = 'menu',
-
-  BELL = 'bell',
-
-  LANGUAGE = 'language',
-
-  TRASH = 'trash',
-
-  HEART = 'heart',
-
-  ELLIPSIS = 'ellipsis',
-}
